# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Account Manager Sync API
  version: 0.2.3
  description: |
    Account Manager 的自托管同步 API，支持端到端加密的 TOTP/2FA 和账号资料管理。

    服务端认证账号并存储密文。
    它从不接收同步密码，也无法解密用户数据。
    请参阅 `docs/security.md` 和 `docs/sync-protocol.md`。
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发环境
...
security:
  - SessionAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: 注册新账号。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 账号已创建。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /v1/auth/login:
    post:
      tags: [auth]
      summary: 认证已有账号。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 会话已建立。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: 吊销当前会话。
      responses:
        '204':
          description: 已注销。
        '401':
          $ref: '#/components/responses/Error'
  /v1/auth/me:
    get:
      tags: [auth]
      summary: 返回已认证用户的身份信息。
      responses:
        '200':
          description: 调用者身份。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        '401':
          $ref: '#/components/responses/Error'
  /v1/auth/password:
    put:
      tags: [auth]
      summary: 修改账号密码。
      description: |
        修改账号密码不会更改同步密码，
        也不会重新加密密钥库。服务端仅更新其
        密码哈希值。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeAccountPasswordRequest'
      responses:
        '204':
          description: 密码已更新。
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'

  # ---------- Devices ----------
  /v1/devices:
    get:
      tags: [devices]
      summary: 列出调用者的设备。
      responses:
        '200':
          description: 设备列表。
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                additionalProperties: false
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [devices]
      summary: 注册新设备并将当前会话绑定到该设备。
      description: |
        客户端生成不透明 `device_id` 并传入 `{id, label}`。
        服务端原子性地创建设备记录并将当前会话绑定到该设备。
        注册成功后该会话即可用于同步操作。响应仅返回 `Device`
        对象，不返回新会话令牌。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: 设备已注册，当前会话已绑定到该设备。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          description: "`device_exists`：该 `device_id` 已被注册；响应不披露设备归属。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    delete:
      tags: [devices]
      summary: 吊销调用者拥有的设备。
      responses:
        '204':
          description: 设备已吊销。
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

  # ---------- Sync ----------
  /v1/sync/vault:
    get:
      tags: [sync]
      summary: 获取已认证用户的密钥库元数据。
      responses:
        '200':
          description: 密钥库元数据。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
        '401':
          $ref: '#/components/responses/Error'
  /v1/sync/vault/envelope:
    put:
      tags: [sync]
      summary: 替换密钥库信封（首次上传或同步密码变更）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutEnvelopeRequest'
      responses:
        '200':
          description: 信封已更新。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          description: "`expected_rev` 过期。返回当前密钥库。"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
  /v1/sync/pull:
    post:
      tags: [sync]
      summary: "拉取 `seq > since_seq` 的增量。"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PullRequest'
      responses:
        '200':
          description: 增量批次。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullResponse'
        '401':
          $ref: '#/components/responses/Error'
  /v1/sync/push:
    post:
      tags: [sync]
      summary: 推送记录级变更，使用乐观并发控制。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRequest'
      responses:
        '200':
          description: 推送已处理；部分记录可能存在冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'
        '401':
          $ref: '#/components/responses/Error'
  /v1/sync/items/{itemId}:
    parameters:
      - $ref: '#/components/parameters/ItemId'
    get:
      tags: [sync]
      summary: 按 ID 获取单条项目记录。
      responses:
        '200':
          description: 项目记录。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/sync/groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [sync]
      summary: 按 ID 获取单条分组记录。
      responses:
        '200':
          description: 分组记录。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/sync/accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [sync]
      summary: 按 ID 获取单条账户元数据和加密密文记录。
      responses:
        '200':
          description: 账户记录。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/sync/relations/{relationId}:
    parameters:
      - $ref: '#/components/parameters/RelationId'
    get:
      tags: [sync]
      summary: 按 ID 获取单条关系记录。
      responses:
        '200':
          description: 关系记录。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relation'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

  # ---------- Admin ----------
  /v1/admin/auth/login:
    post:
      tags: [admin]
      summary: 管理员登录（独立于用户认证）。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 管理员会话已建立。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Error'
  /v1/admin/users:
    get:
      tags: [admin]
      summary: 列出用户（仅非敏感元数据）。
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: 用户分页。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserPage'
        '401':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [admin]
      summary: 获取单个用户（仅非敏感元数据）。
      responses:
        '200':
          description: 某用户的管理员视图。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}/disable:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: [admin]
      summary: 禁用用户（暂停同步；无法解密任何数据）。
      responses:
        '204':
          description: 用户已禁用。
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}/enable:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: [admin]
      summary: 重新启用之前已禁用的用户。
      responses:
        '204':
          description: 用户已启用。
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}/devices:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [admin]
      summary: 列出某用户的已注册设备。
      responses:
        '200':
          description: 设备列表，仅包含非敏感元数据。
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}/accounts:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [admin]
      summary: 列出某用户的账户元数据，不返回 secret_ciphertext。
      responses:
        '200':
          description: 账户元数据列表，不含密文。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccountPage'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}/relations:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [admin]
      summary: 列出某用户的关系元数据，不返回 secret_ciphertext。
      responses:
        '200':
          description: 关系元数据列表，不含密文。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRelationPage'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}/devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/DeviceId'
    delete:
      tags: [admin]
      summary: 吊销某用户的设备。
      responses:
        '204':
          description: 设备已吊销。
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/audit:
    get:
      tags: [admin]
      summary: 列出审计日志条目（不含密文，不含敏感内容）。
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: 审计分页。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPage'
        '401':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    SessionAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
      description: |
        由 `POST /v1/auth/login` 返回的不透明会话令牌。相同的
        认证方案也用于 `/v1/admin/*` 路径的管理员会话；
        服务端在内部执行基于角色的权限范围控制。
  parameters:
    UserId:
      in: path
      name: userId
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    DeviceId:
      in: path
      name: deviceId
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    ItemId:
      in: path
      name: itemId
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    GroupId:
      in: path
      name: groupId
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    AccountId:
      in: path
      name: accountId
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    RelationId:
      in: path
      name: relationId
      required: true
      schema:
        $ref: '#/components/schemas/Id'
  responses:
    Error:
      description: 错误响应。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    Id:
      type: string
      description: 客户端或服务端选择的不透明标识符（UUIDv7 或随机 128 位，base32/hex/ULID）。
      minLength: 8
      maxLength: 64
      example: "01J0Q5K9M9V3SDZQ3K2T7R7B8H"
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 / ISO 8601 格式的 UTC 时间戳。

    # -------- Errors --------
    ApiError:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          description: |
            稳定的机器可读错误代码。认证凭据和会话错误包括：
            - `auth.unauthorized`：缺少或格式错误的认证凭据。
            - `auth.session_expired`：空闲超时或绝对生命周期已到期。
            - `auth.session_invalid`：会话不存在或无效。
            - `auth.session_revoked`：会话绑定的设备缺失或已撤销。
            - `auth.user_disabled`：会话用户已被禁用。
          example: "auth.invalid_credentials"
        message:
          type: string
          description: 人类可读、对最终用户安全的消息。
        details:
          type: object
          additionalProperties: true
          description: 可选的结构化详情。绝不包含密钥。

    # -------- Meta --------
...
    RegisterRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
          description: 账号密码。并非同步密码。
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 1
          maxLength: 256
        device_id:
          $ref: '#/components/schemas/Id'
          description: 可选的客户端选择设备 ID，用于绑定会话。
    ChangeAccountPasswordRequest:
      type: object
      required: [current_password, new_password]
      additionalProperties: false
      properties:
        current_password:
          type: string
          minLength: 1
          maxLength: 256
        new_password:
          type: string
          minLength: 12
          maxLength: 256
    SessionResponse:
      type: object
      required: [token, user]
      additionalProperties: false
      properties:
        token:
          type: string
          description: "不透明会话令牌。以 `Authorization: Bearer <token>` 形式传回。"
        user:
          $ref: '#/components/schemas/UserMe'
        expires_at:
          $ref: '#/components/schemas/Timestamp'
    UserMe:
      type: object
      required: [id, username, role, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        disabled:
          type: boolean
          default: false
        created_at:
          $ref: '#/components/schemas/Timestamp'

    # -------- Devices --------
    Device:
      type: object
      required: [id, label, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        label:
          type: string
          description: 用户选择的人类可读设备标签。
          maxLength: 100
        created_at:
          $ref: '#/components/schemas/Timestamp'
        last_seen_at:
          $ref: '#/components/schemas/Timestamp'
        revoked:
          type: boolean
          default: false
    RegisterDeviceRequest:
      type: object
      required: [id, label]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        label:
          type: string
          maxLength: 100

    # -------- Crypto envelopes --------
...
    VaultEnvelope:
      type: object
      required: [alg, kdf, kdf_salt_b64, wrapped_dek_b64, wrap_iv_b64]
      additionalProperties: false
      properties:
        alg:
          type: string
          description: 用于包裹 DEK 的 AEAD 算法。
          example: "A256GCM"
        kdf:
          type: string
          description: 用于从同步密码派生包裹密钥的 KDF 算法。
          example: "pbkdf2"
        kdf_params:
          $ref: '#/components/schemas/KdfParams'
        kdf_salt_b64:
          $ref: '#/components/schemas/Base64'
        wrapped_dek_b64:
          $ref: '#/components/schemas/Base64'
        wrap_iv_b64:
          $ref: '#/components/schemas/Base64'
...
    Vault:
      type: object
      required: [user_id, seq, envelope_rev, created_at, updated_at]
      additionalProperties: false
      properties:
        user_id:
          $ref: '#/components/schemas/Id'
        seq:
          type: integer
          format: int64
          minimum: 0
          description: 该用户当前的单调同步计数器。
        envelope_rev:
          type: integer
          format: int64
          minimum: 0
          description: 信封记录的乐观并发令牌。
        envelope:
          $ref: '#/components/schemas/VaultEnvelope'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    PutEnvelopeRequest:
      type: object
      required: [envelope, expected_rev]
      additionalProperties: false
      properties:
        envelope:
          $ref: '#/components/schemas/VaultEnvelope'
        expected_rev:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          description: 使用 0 / null 表示"尚无信封"。

    # -------- Records --------
    Item:
      type: object
      required: [id, rev, seq, deleted, updated_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        group_id:
          $ref: '#/components/schemas/Id'
          nullable: true
        rev:
          type: integer
          format: int64
          minimum: 1
        seq:
          type: integer
          format: int64
          minimum: 1
        deleted:
          type: boolean
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        ciphertext:
          $ref: '#/components/schemas/RecordCipher'
          nullable: true
          description: 当 `deleted` 为 true 时为 null。
    Group:
      type: object
      required: [id, rev, seq, deleted, sort_index, updated_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        rev:
          type: integer
          format: int64
          minimum: 1
        seq:
          type: integer
          format: int64
          minimum: 1
        deleted:
          type: boolean
        sort_index:
          type: integer
          format: int64
          description: 明文排序提示。
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        ciphertext:
          $ref: '#/components/schemas/RecordCipher'
          nullable: true
          description: 当 `deleted` 为 true 时为 null。
    Account:
      type: object
      required: [id, rev, seq, deleted, kind, platform, display_name, status, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        rev:
          type: integer
          format: int64
          minimum: 1
        seq:
          type: integer
          format: int64
          minimum: 1
        deleted:
          type: boolean
        kind:
          type: string
          description: 账户类型，例如 login、totp、proxy_profile。
        platform:
          type: string
          description: 平台或服务名。
        display_name:
          type: string
          description: 管理端可见的脱敏显示名称。
        login_identifier:
          type: string
          nullable: true
          description: 脱敏后的登录标识；不得包含完整手机号、完整邮箱或敏感令牌。
        login_identifier_hash:
          type: string
          nullable: true
          description: 客户端计算的登录标识哈希，用于去重/关联。
        status:
          type: string
        tags_json:
          description: JSON 标签数组或对象，仅限非敏感元数据。
          nullable: true
        metadata_json:
          description: JSON 元数据，仅限非敏感字段。
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        secret_ciphertext:
          $ref: '#/components/schemas/RecordCipher'
          nullable: true
          description: 包含密码、TOTP、恢复码、token、cookie、代理认证等敏感内容的加密密文；管理员接口永不返回。
    Relation:
      type: object
      required: [id, rev, seq, deleted, kind, from_kind, from_id, to_kind, to_id, relation_type, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        rev:
          type: integer
          format: int64
          minimum: 1
        seq:
          type: integer
          format: int64
          minimum: 1
        deleted:
          type: boolean
        kind:
          type: string
        from_kind:
          type: string
        from_id:
          $ref: '#/components/schemas/Id'
        to_kind:
          type: string
        to_id:
          $ref: '#/components/schemas/Id'
        relation_type:
          type: string
          description: 与 `kind` 等价的设计文档别名。服务端响应会补齐该字段，方便按账号关系视角阅读。
        from_account_id:
          allOf:
            - $ref: '#/components/schemas/Id'
          nullable: true
          description: 当 `from_kind == "account"` 时等于 `from_id`；其他情况下为空，避免误把非账号端点当账号引用。
        to_account_id:
          allOf:
            - $ref: '#/components/schemas/Id'
          nullable: true
          description: 当 `to_kind == "account"` 时等于 `to_id`；其他情况下为空。
        metadata_json:
          description: JSON 元数据，仅限非敏感字段。
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        secret_ciphertext:
          $ref: '#/components/schemas/RecordCipher'
          nullable: true
          description: 关系相关敏感内容的加密密文；管理员接口永不返回。

    # -------- Sync push/pull --------
    PullRequest:
      type: object
      required: [since_seq]
      additionalProperties: false
      properties:
        since_seq:
          type: integer
          format: int64
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 500
    PullResponse:
      type: object
      required: [items, groups, next_seq, has_more]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        accounts:
          type: array
          description: 账号元数据增量。旧客户端可忽略；未启用账号投影时可能省略。
          items:
            $ref: '#/components/schemas/Account'
        relations:
          type: array
          description: 账号关系元数据增量。旧客户端可忽略；未启用关系投影时可能省略。
          items:
            $ref: '#/components/schemas/Relation'
        next_seq:
          type: integer
          format: int64
          minimum: 0
        has_more:
          type: boolean
    PushItem:
      type: object
      required: [id, deleted, expected_rev]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        group_id:
          $ref: '#/components/schemas/Id'
          nullable: true
        deleted:
          type: boolean
        expected_rev:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          description: Null 表示仅创建。否则必须匹配服务端 rev。
        ciphertext:
          $ref: '#/components/schemas/RecordCipher'
          nullable: true
          description: 当 `deleted` 为 false 时必填。
    PushGroup:
      type: object
      required: [id, deleted, sort_index, expected_rev]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        deleted:
          type: boolean
        sort_index:
          type: integer
          format: int64
        expected_rev:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        ciphertext:
          $ref: '#/components/schemas/RecordCipher'
          nullable: true
...
    PushRequest:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PushItem'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/PushGroup'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/PushAccount'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/PushRelation'
...
    PushResponse:
      type: object
      required: [applied, conflicts, next_seq]
      additionalProperties: false
      properties:
        applied:
          type: array
          items:
            $ref: '#/components/schemas/AppliedRecord'
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/ConflictRecord'
        next_seq:
          type: integer
          format: int64
          minimum: 0

    # -------- Admin --------
    AdminUser:
      type: object
      required: [id, username, role, disabled, device_count, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        disabled:
          type: boolean
        device_count:
          type: integer
          minimum: 0
        last_sync_at:
          $ref: '#/components/schemas/Timestamp'
        ciphertext_bytes:
          type: integer
          format: int64
          minimum: 0
          description: 该用户使用的密文存储总量，用于容量规划。
        created_at:
          $ref: '#/components/schemas/Timestamp'
    AdminUserPage:
      type: object
      required: [users]
      additionalProperties: false
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        next_cursor:
          type: string
          nullable: true
...
    AdminAccountPage:
      type: object
      required: [accounts]
      additionalProperties: false
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AdminAccount'
    AdminRelationPage:
      type: object
      required: [relations]
      additionalProperties: false
      properties:
        relations:
          type: array
          items:
            $ref: '#/components/schemas/AdminRelation'
...
    AuditPage:
      type: object
      required: [entries]
      additionalProperties: false
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/AuditEntry'
        next_cursor:
          type: string
          nullable: true